Factor out duplication state tracking logic to a concern * Also changes the behavior to only clearing the flag after commit
